// packages/sales-common/src/schemas/schema.types.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Decimal: { input: any; output: any; }
  JSON: { input: any; output: any; }
  NonEmptyString: { input: any; output: any; }
  NonNegativeDecimal: { input: any; output: any; }
  NonNegativeInt: { input: any; output: any; }
  PositiveDecimal: { input: any; output: any; }
  PositiveInt: { input: any; output: any; }
  _Any: { input: any; output: any; }
  _FieldSet: { input: any; output: any; }
};

export type BaseOptionsInput = {
  limit?: InputMaybe<Scalars['NonNegativeInt']['input']>;
  offset?: InputMaybe<Scalars['NonNegativeInt']['input']>;
  prefix?: InputMaybe<Scalars['NonEmptyString']['input']>;
};

export type BasicCommonAccount = {
  __typename?: 'BasicCommonAccount';
  bank?: Maybe<BasicCommonBank>;
  id: Scalars['PositiveInt']['output'];
  number?: Maybe<Scalars['String']['output']>;
};

export type BasicCommonBank = {
  __typename?: 'BasicCommonBank';
  bik?: Maybe<Scalars['String']['output']>;
  city?: Maybe<Scalars['String']['output']>;
  correspondent_number?: Maybe<Scalars['String']['output']>;
  id: Scalars['PositiveInt']['output'];
  name: Scalars['String']['output'];
};

export type BasicCommonContractor = {
  __typename?: 'BasicCommonContractor';
  actual_address?: Maybe<Scalars['String']['output']>;
  contacts?: Maybe<Scalars['String']['output']>;
  id: Scalars['PositiveInt']['output'];
  inn?: Maybe<Scalars['String']['output']>;
  is_active?: Maybe<Scalars['Boolean']['output']>;
  kpp?: Maybe<Scalars['String']['output']>;
  legal_address?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  ogrn?: Maybe<Scalars['String']['output']>;
  phone?: Maybe<Scalars['String']['output']>;
  propogated_at?: Maybe<Scalars['DateTime']['output']>;
  reconciliation_link?: Maybe<Scalars['String']['output']>;
  short_name?: Maybe<Scalars['String']['output']>;
};

export type BasicCommonEntity = {
  __typename?: 'BasicCommonEntity';
  db_name: Scalars['String']['output'];
  display_name: Scalars['String']['output'];
  id: Scalars['PositiveInt']['output'];
  inn: Scalars['String']['output'];
  kpp: Scalars['String']['output'];
  name: Scalars['String']['output'];
  ogrn: Scalars['String']['output'];
  short_name: Scalars['String']['output'];
  str_id: Scalars['String']['output'];
};

export type CommonContractor = {
  __typename?: 'CommonContractor';
  accounts?: Maybe<Array<BasicCommonAccount>>;
  contractor: BasicCommonContractor;
};

export type CommonContractors = {
  __typename?: 'CommonContractors';
  contractors: Array<CommonContractor>;
  total_count: Scalars['PositiveInt']['output'];
};

export type CommonEntities = {
  __typename?: 'CommonEntities';
  entities: Array<BasicCommonEntity>;
  total_count: Scalars['PositiveInt']['output'];
};

export type GetCommonContractorInput = {
  id: Scalars['PositiveInt']['input'];
};

export type GetCommonContractorsInput = {
  options?: InputMaybe<BaseOptionsInput>;
};

export type GetCommonEntitiesInput = {
  options?: InputMaybe<BaseOptionsInput>;
};

export type GetCommonEntityInput = {
  id: Scalars['PositiveInt']['input'];
};

export type Query = {
  __typename?: 'Query';
  _service: _Service;
  getCommonContractor: CommonContractor;
  getCommonContractors: CommonContractors;
  getCommonEntities: CommonEntities;
  getCommonEntity: BasicCommonEntity;
};


export type QueryGetCommonContractorArgs = {
  input: GetCommonContractorInput;
};


export type QueryGetCommonContractorsArgs = {
  input?: InputMaybe<GetCommonContractorsInput>;
};


export type QueryGetCommonEntitiesArgs = {
  input?: InputMaybe<GetCommonEntitiesInput>;
};


export type QueryGetCommonEntityArgs = {
  input: GetCommonEntityInput;
};

export type _Service = {
  __typename?: '_Service';
  sdl?: Maybe<Scalars['String']['output']>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;
// ...
export type ResolversParentTypes = {
  BaseOptionsInput: BaseOptionsInput;
  BasicCommonAccount: BasicCommonAccount;
  BasicCommonBank: BasicCommonBank;
  BasicCommonContractor: BasicCommonContractor;
  BasicCommonEntity: BasicCommonEntity;
  Boolean: Scalars['Boolean']['output'];
  CommonContractor: CommonContractor;
  CommonContractors: CommonContractors;
  CommonEntities: CommonEntities;
  DateTime: Scalars['DateTime']['output'];
  Decimal: Scalars['Decimal']['output'];
  GetCommonContractorInput: GetCommonContractorInput;
  GetCommonContractorsInput: GetCommonContractorsInput;
  GetCommonEntitiesInput: GetCommonEntitiesInput;
  GetCommonEntityInput: GetCommonEntityInput;
  JSON: Scalars['JSON']['output'];
  NonEmptyString: Scalars['NonEmptyString']['output'];
  NonNegativeDecimal: Scalars['NonNegativeDecimal']['output'];
  NonNegativeInt: Scalars['NonNegativeInt']['output'];
  PositiveDecimal: Scalars['PositiveDecimal']['output'];
  PositiveInt: Scalars['PositiveInt']['output'];
  Query: {};
  String: Scalars['String']['output'];
  _Any: Scalars['_Any']['output'];
  _FieldSet: Scalars['_FieldSet']['output'];
  _Service: _Service;
};
// ...
export type BasicCommonAccountResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicCommonAccount'] = ResolversParentTypes['BasicCommonAccount']> = {
  bank?: Resolver<Maybe<ResolversTypes['BasicCommonBank']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  number?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicCommonBankResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicCommonBank'] = ResolversParentTypes['BasicCommonBank']> = {
  bik?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  city?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  correspondent_number?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicCommonContractorResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicCommonContractor'] = ResolversParentTypes['BasicCommonContractor']> = {
  actual_address?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  contacts?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  inn?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  is_active?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
  kpp?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  legal_address?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  ogrn?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  phone?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  propogated_at?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
  reconciliation_link?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  short_name?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicCommonEntityResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicCommonEntity'] = ResolversParentTypes['BasicCommonEntity']> = {
  db_name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  display_name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  inn?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  kpp?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  ogrn?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  short_name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  str_id?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type CommonContractorResolvers<ContextType = any, ParentType extends ResolversParentTypes['CommonContractor'] = ResolversParentTypes['CommonContractor']> = {
  accounts?: Resolver<Maybe<Array<ResolversTypes['BasicCommonAccount']>>, ParentType, ContextType>;
  contractor?: Resolver<ResolversTypes['BasicCommonContractor'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type CommonContractorsResolvers<ContextType = any, ParentType extends ResolversParentTypes['CommonContractors'] = ResolversParentTypes['CommonContractors']> = {
  contractors?: Resolver<Array<ResolversTypes['CommonContractor']>, ParentType, ContextType>;
  total_count?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type CommonEntitiesResolvers<ContextType = any, ParentType extends ResolversParentTypes['CommonEntities'] = ResolversParentTypes['CommonEntities']> = {
  entities?: Resolver<Array<ResolversTypes['BasicCommonEntity']>, ParentType, ContextType>;
  total_count?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  _service?: Resolver<ResolversTypes['_Service'], ParentType, ContextType>;
  getCommonContractor?: Resolver<ResolversTypes['CommonContractor'], ParentType, ContextType, RequireFields<QueryGetCommonContractorArgs, 'input'>>;
  getCommonContractors?: Resolver<ResolversTypes['CommonContractors'], ParentType, ContextType, Partial<QueryGetCommonContractorsArgs>>;
  getCommonEntities?: Resolver<ResolversTypes['CommonEntities'], ParentType, ContextType, Partial<QueryGetCommonEntitiesArgs>>;
  getCommonEntity?: Resolver<ResolversTypes['BasicCommonEntity'], ParentType, ContextType, RequireFields<QueryGetCommonEntityArgs, 'input'>>;
};

export interface _AnyScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['_Any'], any> {
  name: '_Any';
}

export interface _FieldSetScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['_FieldSet'], any> {
  name: '_FieldSet';
}

export type _ServiceResolvers<ContextType = any, ParentType extends ResolversParentTypes['_Service'] = ResolversParentTypes['_Service']> = {
  sdl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type Resolvers<ContextType = any> = {
  BasicCommonAccount?: BasicCommonAccountResolvers<ContextType>;
  BasicCommonBank?: BasicCommonBankResolvers<ContextType>;
  BasicCommonContractor?: BasicCommonContractorResolvers<ContextType>;
  BasicCommonEntity?: BasicCommonEntityResolvers<ContextType>;
  CommonContractor?: CommonContractorResolvers<ContextType>;
  CommonContractors?: CommonContractorsResolvers<ContextType>;
  CommonEntities?: CommonEntitiesResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  Decimal?: GraphQLScalarType;
  JSON?: GraphQLScalarType;
  NonEmptyString?: GraphQLScalarType;
  NonNegativeDecimal?: GraphQLScalarType;
  NonNegativeInt?: GraphQLScalarType;
  PositiveDecimal?: GraphQLScalarType;
  PositiveInt?: GraphQLScalarType;
  Query?: QueryResolvers<ContextType>;
  _Any?: GraphQLScalarType;
  _FieldSet?: GraphQLScalarType;
  _Service?: _ServiceResolvers<ContextType>;
};